pyproject.toml
[project]
name = "cvfitengine"
version = "0.1.0"
description = "Role-aligned CV assembly engine"
requires-python = ">=3.10"

dependencies = [
  "pyyaml>=6.0",
  "pydantic>=2.0",
  "jinja2>=3.1",
  "streamlit>=1.33"
]


[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
